Remove forced eval of local filter calls
This parsing of calls to identify bits that should be evaluated locally was quite fragile and only possible due to previously loose tolerances in dbplyr::partial_eval. With the change in dbplyr 2.2.2 (https://github.com/tidyverse/dbplyr/pull/907) this loophole is closed, and exposes the fact that the identification of parts of the call that should be evaluated locally didn't always work. With this change users will now need to explicitly use `local()` to wrap function calls that should be evaluated locally before being translated to CQL and sent to the wfs server.